--- src/main/java/coz/weavon/common/io/PageRequest.java
@@ -1,15 +1,0 @@
-package coz.weavon.common.io;
-
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
-@Getter
-@Setter
-@NoArgsConstructor
-public class PageRequest {
-
-    private Integer pageNo;
-
-    private Integer pageSize;
-}
