+++ src/main/java/coz/weavon/common/response/ErrorResponse.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.io;
+package coz.weavon.common.response;
 
 import lombok.Builder;
 import lombok.Getter;
