--- src/main/java/coz/weavon/common/shared/Adapter.java
@@ -1,19 +1,0 @@
-package coz.weavon.common.shared;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-import org.springframework.core.annotation.AliasFor;
-import org.springframework.stereotype.Component;
-
-@Component
-@Documented
-@Target({ElementType.TYPE})
-@Retention(RetentionPolicy.RUNTIME)
-public @interface Adapter {
-
-    @AliasFor(annotation = Component.class)
-    String value() default "";
-}
