+++ src/main/java/coz/weavon/core/user/application/model/command/UserSearchCommand.java
@@ -1,12 +1,12 @@
 package coz.weavon.core.user.application.model.command;
 
-import coz.weavon.common.shared.Command;
+import coz.weavon.common.command.Command;
 import java.util.List;
-import lombok.Builder;
 import lombok.Getter;
+import lombok.experimental.SuperBuilder;
 
 @Getter
-@Builder
+@SuperBuilder
 public class UserSearchCommand extends Command {
 
     private List<Long> userIds;
