+++ src/main/java/coz/weavon/exception/AuthException.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.exception;
+package coz.weavon.exception;
 
 public class AuthException extends RestException {
