+++ src/main/java/coz/weavon/exception/BusinessException.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.exception;
+package coz.weavon.exception;
 
 public class BusinessException extends RestException {
