+++ src/main/java/coz/weavon/validator/ValidRange.java
@@ -1,4 +1,4 @@
-package coz.weavon.common.validator;
+package coz.weavon.validator;
 
 import jakarta.validation.Constraint;
 import java.lang.annotation.ElementType;
@@ -8,8 +8,12 @@ import java.lang.annotation.Target;
 
 @Target({ElementType.TYPE})
 @Retention(RetentionPolicy.RUNTIME)
-@Constraint(validatedBy = PaginationValidator.class)
-public @interface ValidPagination {
+@Constraint(validatedBy = RangeValidator.class)
+public @interface ValidRange {
 
     boolean required() default true;
+
+    String startFieldName() default "startDate";
+
+    String endFieldName() default "endDate";
 }
